+++ haskell-lsp-types/src/Language/Haskell/LSP/Types/DataTypesJSON.hs
@@ -10,25 +10,7 @@
 {-# LANGUAGE TemplateHaskell            #-}
 {-# LANGUAGE TypeSynonymInstances       #-}
 
-module Language.Haskell.LSP.Types.DataTypesJSON
-    ( module Language.Haskell.LSP.Types.DataTypesJSON
-    , module Language.Haskell.LSP.Types.CodeAction
-    , module Language.Haskell.LSP.Types.Color
-    , module Language.Haskell.LSP.Types.Command
-    , module Language.Haskell.LSP.Types.Completion
-    , module Language.Haskell.LSP.Types.Diagnostic
-    , module Language.Haskell.LSP.Types.DocumentFilter
-    , module Language.Haskell.LSP.Types.FoldingRange
-    , module Language.Haskell.LSP.Types.List
-    , module Language.Haskell.LSP.Types.Location
-    , module Language.Haskell.LSP.Types.MarkupContent
-    , module Language.Haskell.LSP.Types.Message
-    , module Language.Haskell.LSP.Types.Symbol
-    , module Language.Haskell.LSP.Types.TextDocument
-    , module Language.Haskell.LSP.Types.Uri
-    , module Language.Haskell.LSP.Types.WorkspaceEdit
-    , module Language.Haskell.LSP.Types.WorkspaceFolders
-    ) where
+module Language.Haskell.LSP.Types.DataTypesJSON where
 
 import           Control.Applicative
 import qualified Data.Aeson                                 as A
@@ -36,20 +18,14 @@ import           Data.Aeson.TH
 import           Data.Aeson.Types
 import           Data.Text                                  (Text)
 import qualified Data.Text                                  as T
-
-import           Language.Haskell.LSP.Types.ClientCapabilities
-import           Language.Haskell.LSP.Types.CodeAction
-import           Language.Haskell.LSP.Types.Color
 import           Language.Haskell.LSP.Types.Command
-import           Language.Haskell.LSP.Types.Completion
 import           Language.Haskell.LSP.Types.Constants
+import           Language.Haskell.LSP.Types.ClientCapabilities
 import           Language.Haskell.LSP.Types.Diagnostic
 import           Language.Haskell.LSP.Types.DocumentFilter
-import           Language.Haskell.LSP.Types.FoldingRange
 import           Language.Haskell.LSP.Types.List
-import           Language.Haskell.LSP.Types.Location
-import           Language.Haskell.LSP.Types.MarkupContent
 import           Language.Haskell.LSP.Types.Message
+import           Language.Haskell.LSP.Types.Location
 import           Language.Haskell.LSP.Types.Symbol
 import           Language.Haskell.LSP.Types.TextDocument
 import           Language.Haskell.LSP.Types.Uri